added simple test
